Don't setup logger configuration on each notification
It uses a shared instance for each BreezFcmService session, there's no point in updating configuration.
