+++ android/app/src/main/java/com/cBreez/client/BreezFCMService.kt
@@ -5,12 +5,11 @@ import android.app.ActivityManager
 import android.app.KeyguardManager
 import android.os.Process
 import android.os.SystemClock
+import com.cBreez.client.BreezLogger.Companion.configureLogger
 import com.google.android.gms.common.util.PlatformVersion.isAtLeastLollipop
 import com.google.firebase.messaging.FirebaseMessagingService
 import com.google.firebase.messaging.RemoteMessage
-import io.flutter.util.PathUtils
 import org.tinylog.kotlin.Logger
-import java.io.File
 
 @SuppressLint("MissingFirebaseInstanceTokenRefresh")
 class BreezFcmService : FirebaseMessagingService() {
@@ -20,7 +19,7 @@ class BreezFcmService : FirebaseMessagingService() {
 
     override fun onMessageReceived(remoteMessage: RemoteMessage) {
         super.onMessageReceived(remoteMessage)
-        configureLogger()
+        configureLogger(applicationContext)
         // Only handle remote messages if app is in the background
         if (isAppForeground()) {
             Logger.tag(TAG).info { "App is in the foreground." }
@@ -80,18 +79,5 @@ class BreezFcmService : FirebaseMessagingService() {
             }
         }
         return false
-    }
-
-    private fun configureLogger() {
-        val loggingDir: File = File(
-            PathUtils.getDataDirectory(applicationContext),
-            "/logs/android-extension/",
-        ).also { it.mkdirs() }
-
-        System.setProperty("tinylog.directory", loggingDir.absolutePath)
-        System.setProperty("tinylog.timestamp", System.currentTimeMillis().toString())
-
-        Logger.tag(TAG).info { "Starting ${BuildConfig.APPLICATION_ID}..." }
-        Logger.tag(TAG).info { "Logs directory: '$loggingDir'" }
     }
 }
